feat!: import Traefik logs into GoAccess via one-off Compose task
As of https://github.com/mobidata-bw/ipl-ansible/commit/b63d284218a3878bf6b9b386c6cf36195f5eeedb, we rotate the Traefik/ingress access logs using logrotate.

Because the latter runs externally and rotates the log at an arbitrary point in time, the 60-seconds import loop that is currently implemented might miss <=60s worth of logs, because it a) only runs on the current log file (and not rotated ones), and b) its runs are not coordinated with the log rotations.

This commit changes the GoAccess import to allow (and require!) triggering it externally, so that the logrotate setup can do this.