chore: gtfs-api: don't listen on DB channel for schema reloading
PostgREST uses a separate long-running (separate) DB connection to listen on a pub/sub channel for instructions to reload its schema (cache). By default, pgbouncer *does not* force-close idle connections on reconfigurations.

Although we recently configured pgbouncer to do so, we don't need this on-demand schema reloading feature anyways, so we turn it off.